split translationsTab component into manageable pieces
